+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Fix: new message dialog tab order (!965)
 - Fix: automatically open talking stick tab for moderator on start (!960)
+- Fix: make clear search button focusable with tab key (!961)
 
 ### Stability Improvements